WiP abstracted ai handling into a base class, ui handler and ollama specific integration class
